% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compartment_dynamics.R
\name{compartment_sizes_for_group}
\alias{compartment_sizes_for_group}
\title{Get the SEIR compartment sizes for a group}
\usage{
compartment_sizes_for_group(group, max_time, frequency = 1)
}
\arguments{
\item{group}{A Group object}

\item{max_time}{The last time to get compartment sizes for. Compartment sizes will be retrieved for times from 0 to \code{max_time}}

\item{frequency}{Integer indicating how frequently the compartment sizes should be measured}
}
\value{
A list with: time, group id, and SEIR sizes
}
\description{
Get the SEIR compartment sizes for a group
}
