% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compartment_dynamics.R
\name{compartment_sizes_for_groups}
\alias{compartment_sizes_for_groups}
\title{Get the SEIR compartment sizes for set of groups}
\usage{
compartment_sizes_for_groups(groups, max_time, frequency = 1)
}
\arguments{
\item{groups}{A vector of group objects}

\item{max_time}{The last time to get compartment sizes for. Compartment sizes will be retrieved for times from 0 to \code{max_time}}

\item{frequency}{Integer indicating how frequently the compartment sizes should be measured}
}
\value{
A data.frame with the columns: time, group id, and SEIR sizes and \code{length(groups)*time}
}
\description{
Get the SEIR compartment sizes for set of groups
}
