% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-genome_functions.R
\name{genome_to_characters}
\alias{genome_to_characters}
\title{Convert a genome represeneted by integers to a genome represented by characters}
\usage{
genome_to_characters(int_genome, dna = TRUE)
}
\arguments{
\item{int_genome}{A integer vector compose of 1, 2, 3, & 4}

\item{dna}{A bool indicating if the genome is DNA or RNA based}
}
\value{
A character vector
}
\description{
Convert a genome represeneted by integers to a genome represented by characters
}
