% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-genome_functions.R
\name{genome_to_integers}
\alias{genome_to_integers}
\title{Convert a genome represeneted by characters to a genome represented by integers}
\usage{
genome_to_integers(char_genome, dna = TRUE)
}
\arguments{
\item{char_genome}{A character vector composed of "A", "C", "G" & "T" or "U"}

\item{dna}{A bool indicating if the genome is DNA or RNA based}
}
\value{
A integer vector
}
\description{
Convert a genome represeneted by characters to a genome represented by integers
}
