% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-host.R
\name{Host}
\alias{Host}
\title{Host}
\description{
An R6 class representing a host
}
\examples{
ref_strain <- ReferenceStrain$new(name = 'ref_strain')
grp <- Group$new(id = 1, ref_strain = ref_strain)
Host$new(id = 2, group = grp)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{returns a list of data: host id and group id, infector id and group id, exposure time, infectious time, and recovery time}

\item{\code{exposure_time}}{The time at which this host was exposed (Returns NA if still susceptible)}

\item{\code{freq}}{The frequencies of each \code{Strain} object in \code{strains}}

\item{\code{id}}{The id of this host}

\item{\code{infectious_time}}{The time at which this host became infectious (Returns NA if still susceptible)}

\item{\code{infectees}}{A vector of \code{Host} objects that this \code{host} infected}

\item{\code{infector}}{The \code{Host} object that infected this \code{host}}

\item{\code{is_index}}{Logical indicating if this \code{Host} was an index case}

\item{\code{group}}{The \code{Group} object that this host belongs to}

\item{\code{realisation_time}}{The time at which this \code{Host} last realised itself}

\item{\code{recovery_time}}{The time at which this host recovered (Returns NA if still susceptible)}

\item{\code{sample_freq}}{The \code{sample_freq} specified in this \code{Host}'s \code{Group} object}

\item{\code{sample_time}}{The next time at which this host is due sampling (Returns NA if still susceptible)}

\item{\code{sample_schedule}}{The \code{sample_schedule} specified in this \code{Host}'s \code{Group} object}

\item{\code{strains}}{A vector of \code{Strain} objects that this \code{Host} is carrying}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Host-new}{\code{Host$new()}}
\item \href{#method-Host-contract_strains}{\code{Host$contract_strains()}}
\item \href{#method-Host-infect}{\code{Host$infect()}}
\item \href{#method-Host-is_exposed}{\code{Host$is_exposed()}}
\item \href{#method-Host-is_infectious}{\code{Host$is_infectious()}}
\item \href{#method-Host-is_recovered}{\code{Host$is_recovered()}}
\item \href{#method-Host-is_sampling_due}{\code{Host$is_sampling_due()}}
\item \href{#method-Host-is_susceptible}{\code{Host$is_susceptible()}}
\item \href{#method-Host-prepare_for_infection}{\code{Host$prepare_for_infection()}}
\item \href{#method-Host-print}{\code{Host$print()}}
\item \href{#method-Host-realise}{\code{Host$realise()}}
\item \href{#method-Host-update_sample_time}{\code{Host$update_sample_time()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-new"></a>}}
\if{latex}{\out{\hypertarget{method-Host-new}{}}}
\subsection{Method \code{new()}}{
Create a new host object. Generally, a user won't be required to do this explicitly. It will be handled by a \code{Group} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$new(id, group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{The id of this host. Must be unique within each group, otherwise, this will lead to conflicting lab results.}

\item{\code{group}}{The Group object that this host belongs to (or a R6Class object that inherits from Group).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Host} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-contract_strains"></a>}}
\if{latex}{\out{\hypertarget{method-Host-contract_strains}{}}}
\subsection{Method \code{contract_strains()}}{
Have host contract strains of a pathogen from an infector. The infector chooses which strains are passed on. Note that the S(E)IR times for this host should have already beeen decided for this host by calling prepare_for_infection()
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$contract_strains(infector, strains, freq)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{infector}}{The \code{Host} object that is infecting this \code{host}}

\item{\code{strains}}{A list of \code{Strain} Objects}

\item{\code{freq}}{The frequencies of each \code{Strain} object in \code{strains}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns this host
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-infect"></a>}}
\if{latex}{\out{\hypertarget{method-Host-infect}{}}}
\subsection{Method \code{infect()}}{
Have host infect an infectee
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$infect(infectee, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{infectee}}{The \code{Host} object that is being infected by this \code{host}. The infectee has to have been prepared for infection prior to this function being called}

\item{\code{time}}{The time of the infection}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns this \code{Host} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-is_exposed"></a>}}
\if{latex}{\out{\hypertarget{method-Host-is_exposed}{}}}
\subsection{Method \code{is_exposed()}}{
Test if host is exposed at a given time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$is_exposed(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time being queried}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical indicating if the host was exposed at \code{time}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-is_infectious"></a>}}
\if{latex}{\out{\hypertarget{method-Host-is_infectious}{}}}
\subsection{Method \code{is_infectious()}}{
Test if host is infectious at a given time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$is_infectious(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time being queried}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical indicating if the host was infectious at \code{time}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-is_recovered"></a>}}
\if{latex}{\out{\hypertarget{method-Host-is_recovered}{}}}
\subsection{Method \code{is_recovered()}}{
Test if host is recovered at a given time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$is_recovered(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time being queried}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-is_sampling_due"></a>}}
\if{latex}{\out{\hypertarget{method-Host-is_sampling_due}{}}}
\subsection{Method \code{is_sampling_due()}}{
Test if host is due sampling at a given time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$is_sampling_due(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time being queried}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-is_susceptible"></a>}}
\if{latex}{\out{\hypertarget{method-Host-is_susceptible}{}}}
\subsection{Method \code{is_susceptible()}}{
Test if host is susceptible at a given time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$is_susceptible(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time being queried}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Logical indicating if the host was susceptible at \code{time}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-prepare_for_infection"></a>}}
\if{latex}{\out{\hypertarget{method-Host-prepare_for_infection}{}}}
\subsection{Method \code{prepare_for_infection()}}{
Prepare this host for an infection by deciding the times at which they enter the infectious and recovered compartments and when they are due to be sampled. Finally the host returns an interval which we use to find a their infector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$prepare_for_infection(time, initial = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time of infection}

\item{\code{initial}}{If this is an index case (index cases are placed straight in the infectious compartment)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Time interval that will be used to find a potential infector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-print"></a>}}
\if{latex}{\out{\hypertarget{method-Host-print}{}}}
\subsection{Method \code{print()}}{
Print a description of the \code{Host} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. Description is printed to the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-realise"></a>}}
\if{latex}{\out{\hypertarget{method-Host-realise}{}}}
\subsection{Method \code{realise()}}{
Realise changes in this \code{Host} object: This involves updating the  realisation time and deciding growth of pathogen population (if we implement it)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$realise(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The time of infection}

\item{\code{initial}}{If this is an index case (index cases are placed straight in the infectious compartment)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns this \code{Host} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Host-update_sample_time"></a>}}
\if{latex}{\out{\hypertarget{method-Host-update_sample_time}{}}}
\subsection{Method \code{update_sample_time()}}{
Set a new sampling time in the future. It will be set to the current sampling time plus the \code{sample_freq}. If this new time is after the \code{Host} recovers then it is set to infinity. If \code{sample_schedule} is random nothing will happen.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Host$update_sample_time()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Returns this \code{Host} object
}
}
}
