% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-helper_functions.R
\name{is_ancestor_r6class}
\alias{is_ancestor_r6class}
\title{Function to recursively test if class_1 is the same class as or a ancestor class of class_2}
\usage{
is_ancestor_r6class(class_1, class_2)
}
\arguments{
\item{class_1, class_2}{R6ClassGenerator objects}
}
\value{
A boolean
}
\description{
Function to recursively test if class_1 is the same class as or a ancestor class of class_2
}
