% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-lab.R
\name{Lab}
\alias{Lab}
\title{Lab}
\description{
An R6 class representing a lab. A lab handles taking WGSequences from hosts that are due to be sampled. A user will need to create this object but will usually delegate it's control to either a \code{Group} or \code{Population} object
}
\examples{
#Lab$new()


## ------------------------------------------------
## Method `Lab$sample_hosts`
## ------------------------------------------------

ref_strain <- ReferenceStrain$new("ref_strain", g_len = 100)
grp <- Group$new(
    1, ref_strain,
    init_inf = 6,
    inc_shape = 0,
    sample_schedule = "calendar", sample_freq = 1
)
lab <- Lab$new()
lab$sample_hosts(grp$infectious_hosts(1), 1)

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_wgs}}{The number of whole genome sequences collected so far}

\item{\code{wg_sequences}}{A vector of all whole gnome sequences collected so far}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Lab-new}{\code{Lab$new()}}
\item \href{#method-Lab-print}{\code{Lab$print()}}
\item \href{#method-Lab-fasta_df}{\code{Lab$fasta_df()}}
\item \href{#method-Lab-hostdata_df}{\code{Lab$hostdata_df()}}
\item \href{#method-Lab-sample_hosts}{\code{Lab$sample_hosts()}}
\item \href{#method-Lab-metadata_df}{\code{Lab$metadata_df()}}
\item \href{#method-Lab-save}{\code{Lab$save()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lab-new"></a>}}
\if{latex}{\out{\hypertarget{method-Lab-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Lab} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lab$new(wgs_class = WGSequence)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{wgs_class}}{The class used to create whole genome sequences. Default is \code{WGSequence} though any R6Class that inherits from \code{WGSequence} can be used here}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Lab} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lab-print"></a>}}
\if{latex}{\out{\hypertarget{method-Lab-print}{}}}
\subsection{Method \code{print()}}{
Print a description of this \code{Lab} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lab$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. Description is printed to the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lab-fasta_df"></a>}}
\if{latex}{\out{\hypertarget{method-Lab-fasta_df}{}}}
\subsection{Method \code{fasta_df()}}{
Build a dataframe of sequence names and whole genome sequences for saving to a fasta file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lab$fasta_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lab-hostdata_df"></a>}}
\if{latex}{\out{\hypertarget{method-Lab-hostdata_df}{}}}
\subsection{Method \code{hostdata_df()}}{
Build a dataframe with details about hosts that have been sampled
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lab$hostdata_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lab-sample_hosts"></a>}}
\if{latex}{\out{\hypertarget{method-Lab-sample_hosts}{}}}
\subsection{Method \code{sample_hosts()}}{
Collect samples from \code{host}s
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lab$sample_hosts(hosts, time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hosts}}{A vector of \code{Host} objects to be sampled (or R6Classes that inherit from \code{Host})}

\item{\code{time}}{The time that samples are being collected}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This \code{Lab} object
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ref_strain <- ReferenceStrain$new("ref_strain", g_len = 100)
grp <- Group$new(
    1, ref_strain,
    init_inf = 6,
    inc_shape = 0,
    sample_schedule = "calendar", sample_freq = 1
)
lab <- Lab$new()
lab$sample_hosts(grp$infectious_hosts(1), 1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lab-metadata_df"></a>}}
\if{latex}{\out{\hypertarget{method-Lab-metadata_df}{}}}
\subsection{Method \code{metadata_df()}}{
Build a dataframe with metadata for the whole genome sequences
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lab$metadata_df()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A dataframe
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Lab-save"></a>}}
\if{latex}{\out{\hypertarget{method-Lab-save}{}}}
\subsection{Method \code{save()}}{
Saves host data and sample metadata as csvs and genomes as a fasta file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Lab$save(directory, stem)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{directory}}{The target directory (the directory will be created if it doesn't exist, no warnings will be shown)}

\item{\code{stem}}{the extensionless form of the file name}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This function creates three files:
\itemize{
\item \verb{<directory>/<stem>_hostdata.csv} and
\item \verb{<directory>/<stem>_sample_metadata.csv} and
\item \verb{<directory>/<stem>_samples.fasta}
}
}

\subsection{Returns}{
A dataframe
}
}
}
