% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-outbreak.R
\name{Outbreak}
\alias{Outbreak}
\title{Outbreak}
\description{
An R6 class representing an outbreak. Used to simulate an outbreak in several populations (or groups)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inf_rates}}{An matrix with the same number of rows and columns as there is populations. The entry in the \code{i}th row and \code{j}the column is how much infectious pressure an individual in population \code{i} exerts on an individual in population \code{j}}

\item{\code{lab}}{The \code{Lab} object (or an R6Class that inherits from \code{Lab}) that takes samples from \code{populations}}

\item{\code{populations}}{A vector of \code{Population} objects (or R6Classes that inherit from \code{Population})}

\item{\code{time}}{An integer representing the current time in the outbreak}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{exposed_sizes}}{The number of hosts in the exposed compartment in each population}

\item{\code{hosts_due_for_sampling}}{A vector of hosts from all populations that are due sampling at the current time}

\item{\code{is_outbreak_active}}{Logical indicating if there is any host in any population that is still in the exposed or infectious compartment}

\item{\code{infectious_sizes}}{The number of hosts in the infectious compartment in each population}

\item{\code{recovered_sizes}}{The number of hosts in the recovered compartment in each population}

\item{\code{susceptible_sizes}}{The number of hosts in the susceptible compartment in each population}

\item{\code{num_wgs}}{The number of whole genome sequences collected so far}

\item{\code{outbreak_size}}{The number of non_susceptible hosts across all populations}

\item{\code{population_outbreak_sizes}}{The number of non_susceptible hosts in each population}

\item{\code{population_sizes}}{The size of each population}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Outbreak-new}{\code{Outbreak$new()}}
\item \href{#method-Outbreak-get_infectees}{\code{Outbreak$get_infectees()}}
\item \href{#method-Outbreak-get_infectors}{\code{Outbreak$get_infectors()}}
\item \href{#method-Outbreak-get_population_of_infectors}{\code{Outbreak$get_population_of_infectors()}}
\item \href{#method-Outbreak-get_num_infectees}{\code{Outbreak$get_num_infectees()}}
\item \href{#method-Outbreak-infect}{\code{Outbreak$infect()}}
\item \href{#method-Outbreak-print}{\code{Outbreak$print()}}
\item \href{#method-Outbreak-run_simulation}{\code{Outbreak$run_simulation()}}
\item \href{#method-Outbreak-sample_populations}{\code{Outbreak$sample_populations()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-new"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Outbreak} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$new(populations, inf_rates, lab)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{populations}}{A vector of \code{Population} objects (or R6Classes that inherit from \code{Population})}

\item{\code{inf_rates}}{An matrix with the same number of rows and columns as there is populations. The entry in the \code{i}th row and \code{j}the column is how much infectious pressure an individual in population \code{i} exerts on an individual in population \code{j}}

\item{\code{lab}}{A \code{Lab} object (or R6Class that inherits from \code{Lab})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Outbreak} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-get_infectees"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-get_infectees}{}}}
\subsection{Method \code{get_infectees()}}{
Select \code{host}s from the susceptible hosts in each \code{populations} by calling \code{get_infectees()} for each
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$get_infectees(num_infectees)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_infectees}}{An vector of integers indicating how many infectees to select from each population}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of vectors of \code{Host} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-get_infectors"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-get_infectors}{}}}
\subsection{Method \code{get_infectors()}}{
Select infectors from the infectious hosts in each population in \code{populations} by calling \code{get_infectors()} for each
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$get_infectors(population_of_infectors, infector_intervals)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population_of_infectors}}{A vector of  the population id of each infector.}

\item{\code{infector_intervals}}{A vector of integers indicating how long ago each an infector was infected / infectious (depends on the \code{find_infector_method} in each population). If \code{find_infector_method="random"} then it's a vector of zeros.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of \code{Host} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-get_population_of_infectors"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-get_population_of_infectors}{}}}
\subsection{Method \code{get_population_of_infectors()}}{
Select the population that each infector comes from
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$get_population_of_infectors(num_infectees, fois_mat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_infectees}}{An vector of integers indicating how many infectees to select from each population}

\item{\code{fois_mat}}{A force of infection matrix of the same shape and form as inf_rates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of vectors of population ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-get_num_infectees"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-get_num_infectees}{}}}
\subsection{Method \code{get_num_infectees()}}{
Determine number of infectees in each population
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$get_num_infectees(fois)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fois}}{A vector of the force of infection experienced by each population}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An vector of integers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-infect"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-infect}{}}}
\subsection{Method \code{infect()}}{
Trigger an infection cycle in all \code{populations}. All new infectees (if any) are selected, their infectors are selected and infections of hosts are carried out.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$infect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
This \code{Outbreak} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-print"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-print}{}}}
\subsection{Method \code{print()}}{
Print a description of this \code{Outbreak} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$print()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. Description is printed to the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-run_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-run_simulation}{}}}
\subsection{Method \code{run_simulation()}}{
Run this simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$run_simulation(feedback = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feedback}}{How frequently to print population summary to the console (default 10)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This \code{Outbreak} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Outbreak-sample_populations"></a>}}
\if{latex}{\out{\hypertarget{method-Outbreak-sample_populations}{}}}
\subsection{Method \code{sample_populations()}}{
Collect samples from \code{host}s in every population that are due for sampling
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Outbreak$sample_populations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
This \code{Outbreak} object
}
}
}
