% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-compartment_dynamics.R
\name{plot_compartment_dynamics}
\alias{plot_compartment_dynamics}
\title{Plot S(E)IR compartment dynamics for disease outbreak}
\usage{
plot_compartment_dynamics(
  x,
  frequency = 1,
  combine = FALSE,
  population_labels = NULL
)
}
\arguments{
\item{x}{An Outbreak or Population object that has had it's outbreak simulated}

\item{frequency}{Integer indicating how frequently the compartment sizes should be measured}

\item{combine}{A boolean indicating if compartment sizes for different populations should be combined (only has an effect if an Outbreak object with multiple populations is provided)}

\item{population_labels}{A vector of strings used to label multiple populations. Should be in the same order as the populations in the Outbreak object. (similarly to combine, this only has an effect if an outbreak with multiple populations is provided)}
}
\value{
A ggplot object
}
\description{
Plot S(E)IR compartment dynamics for disease outbreak
}
\examples{
# An SIR model outbreak in a single homogenous population
set.seed(1)
ref_strain <- ReferenceStrain$new("ref_strain")
population <- Population$new(1, ref_strain, inf_rate = 0.75, inc_shape = 0)
lab_pop <- Lab$new()
population$run_simulation(lab_pop)
plot_compartment_dynamics(population)

# An SEIR model outbreak with two populations
set.seed(1)
ref_strain <- ReferenceStrain$new("ref_strain")
outbreak <- Outbreak$new(
    c(
        Population$new(1, ref_strain, init_inf = 5, init_sus = 95),
        Population$new(2, ref_strain, init_inf = 0, init_sus = 100)
    ),
    matrix(c(0.75, 0.5, 0.25, 0.1), ncol = 2),
    Lab$new()
)
outbreak$run_simulation()
# plot each population separately
plot_compartment_dynamics(outbreak, population_labels = c("Senior", "Adult"))
# plot populations combined
plot_compartment_dynamics(outbreak, combine = TRUE)

}
