% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-reference_strain.R
\name{ReferenceStrain}
\alias{ReferenceStrain}
\title{ReferenceStrain}
\description{
An R6 class representing a reference strain for a pathogen. The purpose of this class is to act as a reference for a Strain's genome. It also is where the mutation rate for Strains is defined.
}
\examples{
ReferenceStrain$new(name = "ref_strain", g_len = 1000, mut_rate = 0.00001)
ReferenceStrain$new(name = "ref_strain", genome = c("A", "C", "G", "T"))
ReferenceStrain$new(name = "ref_strain", genome = c("A", "C", "G", "U"), dna = FALSE)

## ------------------------------------------------
## Method `ReferenceStrain$nucleotides_at_loci`
## ------------------------------------------------

ref_strain <- ReferenceStrain$new(name = "ref_strain")
ref_strain$nucleotides_at_loci(c(3, 10, 15))
ref_strain$nucleotides_at_loci(c(3, 10, 15), TRUE)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of this reference strain}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{g_len}}{The length of the reference strain's genome}

\item{\code{is_dna}}{Bool indicating if the genome is DNA based ('FALSE' indicates it is RNA based)}

\item{\code{mut_rate}}{The probability of a single nucleotide polymorphism per unit time.
Mutations follow a Jukes-Cantor model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ReferenceStrain-new}{\code{ReferenceStrain$new()}}
\item \href{#method-ReferenceStrain-genome}{\code{ReferenceStrain$genome()}}
\item \href{#method-ReferenceStrain-nucleotides_at_loci}{\code{ReferenceStrain$nucleotides_at_loci()}}
\item \href{#method-ReferenceStrain-print}{\code{ReferenceStrain$print()}}
\item \href{#method-ReferenceStrain-clone}{\code{ReferenceStrain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceStrain-new"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceStrain-new}{}}}
\subsection{Method \code{new()}}{
Create a new reference strain object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceStrain$new(
  name,
  genome = NULL,
  g_len = 1000L,
  mut_rate = 8e-06,
  dna = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of this reference strain.}

\item{\code{genome}}{A character vector composed of "A", "C", "G", & "T" or "U".
Can be excluded if a randomly generated genome is acceptable.}

\item{\code{g_len}}{Length of randomly generate genome. Not required if a genoms is provided}

\item{\code{mut_rate}}{The probability of a single nucleotide polymorphism per unit time.
Mutations follow a Jukes-Cantor model.}

\item{\code{dna}}{Bool indicating if the genome is DNA or RNA based.
'FALSE' indicates it's RNA based.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{ReferenceStrain} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceStrain-genome"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceStrain-genome}{}}}
\subsection{Method \code{genome()}}{
Get reference strain's genome.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceStrain$genome(as_character = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{as_character}}{Logical indicating if the genome should be returned as a character or integer vector.
Default is TRUE.
'FALSE' returns an integer vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character or integer vector. If an integer vector is specified then 1 = "A", 2 = "C", 3 = "G", and 4 = "T" if genome is DNA based or "U" if genome is RNA based.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceStrain-nucleotides_at_loci"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceStrain-nucleotides_at_loci}{}}}
\subsection{Method \code{nucleotides_at_loci()}}{
Get the nucleotides at the specified loci
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceStrain$nucleotides_at_loci(loci, as_character = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{loci}}{A integer vector specifiying the loci in the genome that you want the nucleotides for.}

\item{\code{as_character}}{Logical indicating if the nucleotides should be returned as a character or integer vector.
Default is FALSE.
'FALSE' returns an integer vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An integer vector representation of the desired nucleotides
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ref_strain <- ReferenceStrain$new(name = "ref_strain")
ref_strain$nucleotides_at_loci(c(3, 10, 15))
ref_strain$nucleotides_at_loci(c(3, 10, 15), TRUE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceStrain-print"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceStrain-print}{}}}
\subsection{Method \code{print()}}{
Print a description of the ReferenceStrain object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceStrain$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. Description is printed to the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ReferenceStrain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ReferenceStrain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ReferenceStrain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
