% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-strain.R
\name{Strain}
\alias{Strain}
\title{Strain}
\description{
An R6 class representing a strain of a pathogen.
}
\examples{
ref_strain <- ReferenceStrain$new(name = "ref_strain")
Strain$new(ref_strain)
Strain$new(ref_strain, time = 3)
Strain$new(ref_strain, distance = 4)

## ------------------------------------------------
## Method `Strain$genome`
## ------------------------------------------------

ref_strain <- ReferenceStrain$new(name = "ref_strain")
strain <- Strain$new(ref_strain)
strain$genome()
strain$genome(as_character = FALSE)

## ------------------------------------------------
## Method `Strain$reproduce`
## ------------------------------------------------

ref_strain <- ReferenceStrain$new(name = "ref_strain")
strain <- Strain$new(ref_strain)
strain$reproduce(time = 12)
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ancestor}}{This strain's immediate ancestor}

\item{\code{descendents}}{A vector of this strain's immediate descendents}

\item{\code{is_dna}}{A logical indicating if this strain has a DNA or RNA based genome (TRUE and FALSE respectively)}

\item{\code{loci}}{A vector of loci where single nucleotide polymorphisms have occured (unsorted)}

\item{\code{nucleotides}}{The nucleotides that are present at the \code{loci} with single nucleotide polymorphisms}

\item{\code{realisation_time}}{The most recent realisation time for this strain}

\item{\code{ref_strain}}{The ReferenceStrain object for this strain}

\item{\code{spawn_time}}{The time that this strain was spawned}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Strain-new}{\code{Strain$new()}}
\item \href{#method-Strain-genome}{\code{Strain$genome()}}
\item \href{#method-Strain-print}{\code{Strain$print()}}
\item \href{#method-Strain-realise}{\code{Strain$realise()}}
\item \href{#method-Strain-reproduce}{\code{Strain$reproduce()}}
\item \href{#method-Strain-clone}{\code{Strain$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strain-new"></a>}}
\if{latex}{\out{\hypertarget{method-Strain-new}{}}}
\subsection{Method \code{new()}}{
Create a new strain object. Generally only used when creating strains for index cases, because of this it's unlikely that a user will create these explicitly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strain$new(ref_strain, time = 0L, distance = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ref_strain}}{A ReferenceStrain object (or a R6Class object that inherits from ReferenceStrain)}

\item{\code{time}}{The time that this strain is 'spawned'}

\item{\code{distance}}{This strains genomic distance from 'ref_strain'. Genomic distance is defined as the number of single nucleotide polymorphisms of this strain compared to it's reference strain.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new 'Strain' object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strain-genome"></a>}}
\if{latex}{\out{\hypertarget{method-Strain-genome}{}}}
\subsection{Method \code{genome()}}{
Get strain's genome
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strain$genome(as_character = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{as_character}}{Logical indicating if the genome should be returned as a character or integer vector.
Default is TRUE.
'FALSE' returns an integer vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character or integer vector. If an integer vector is specified then 1 = "A", 2 = "C", 3 = "G", and 4 = "T" if genome is DNA based or "U" if genome is RNA based.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ref_strain <- ReferenceStrain$new(name = "ref_strain")
strain <- Strain$new(ref_strain)
strain$genome()
strain$genome(as_character = FALSE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strain-print"></a>}}
\if{latex}{\out{\hypertarget{method-Strain-print}{}}}
\subsection{Method \code{print()}}{
Print a description of the \code{ReferenceStrain} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strain$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. Description is printed to the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strain-realise"></a>}}
\if{latex}{\out{\hypertarget{method-Strain-realise}{}}}
\subsection{Method \code{realise()}}{
Realise changes that have occured (ie mutations) that have occured since the last realisation time
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strain$realise(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The current time}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE: The motivation for this is that we can call this function for many strains with vapply if we know what the return value will be, so we've made it something small
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strain-reproduce"></a>}}
\if{latex}{\out{\hypertarget{method-Strain-reproduce}{}}}
\subsection{Method \code{reproduce()}}{
Trigger this strain to reproduce and spawn a descendent
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strain$reproduce(time)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{The current time}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This object's descendent
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{ref_strain <- ReferenceStrain$new(name = "ref_strain")
strain <- Strain$new(ref_strain)
strain$reproduce(time = 12)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strain-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Strain-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strain$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
