% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-transmission_tree.R
\name{transmission_tree_nodes}
\alias{transmission_tree_nodes}
\title{Get a dataframe of all nodes in an outbreak transmission tree}
\usage{
transmission_tree_nodes(x)
}
\arguments{
\item{x}{A Population or Group object that has had it's outbreak simulated}
}
\value{
A dataframe
}
\description{
Get a dataframe of all nodes in an outbreak transmission tree
}
\examples{
set.seed(1)
ref_strain <- ReferenceStrain$new("ref_strain")
group <- Group$new(
     1, 
     ref_strain, 
     inf_rate = 0.75, 
     find_infector_method = "serial", 
     si_shape = 6, si_rate = 1
)
group$run_simulation(Lab$new())
nodes <- transmission_tree_nodes(group)

}
