% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-wgsequence.R
\name{WGSequence}
\alias{WGSequence}
\title{WGSequence}
\description{
An R6 class representing a whole genome sequence test. Users are not expected to interact directly with this class but it is made available so that a user can subclass it if desired.
}
\examples{
ref_strain <- ReferenceStrain$new("ref_strain")
grp <- Group$new(1, ref_strain)
host <- grp$infectious_hosts()[[1]]
strain <- host$strains[[1]]
wgs <- WGSequence$new(1, strain, host)

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{host}}{The \code{Host} object carrying the \code{Strain} object that this sequence was taken from}

\item{\code{loci}}{A vector of loci in \code{strain}'s \code{ReferenceStrain} genome where single nucleotide polymorphisms had occured at time of sampling (unsorted)}

\item{\code{is_dna}}{A logical indicating if this sequence is DNA or RNA based (TRUE and FALSE respectively)}

\item{\code{metadata}}{Returns a named list of data about this sequence: \code{name}, \code{host}'s group id, \code{hosts}'s id, \code{ref_strain}'s name, \code{collection_time}, and \code{sample_number}}

\item{\code{name}}{String in the form \verb{<group_id>_<host_id>_<time>_<sample_number>} padded to have up to 1, 3, 3, and 1 leading zeros respectively}

\item{\code{nucleotides}}{The nucleotides that are present at the \code{loci} with single nucleotide polymorphisms}

\item{\code{ref_strain}}{The \code{ReferenceStrain} object for \code{strain}}

\item{\code{sample_num}}{Integer representing this sample's number for \code{host} at \code{time}}

\item{\code{strain}}{\code{Strain} object that was sequenced}

\item{\code{time}}{When this sequence was recorded}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WGSequence-new}{\code{WGSequence$new()}}
\item \href{#method-WGSequence-genome}{\code{WGSequence$genome()}}
\item \href{#method-WGSequence-print}{\code{WGSequence$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WGSequence-new"></a>}}
\if{latex}{\out{\hypertarget{method-WGSequence-new}{}}}
\subsection{Method \code{new()}}{
Create a new wgsequence object. Generally, a user won't be required to do this explicitly. It will be handled by a \code{Lab} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WGSequence$new(sample_num, strain, host)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sample_num}}{Integer representing this sample's number for this \code{host} at this time (time is retrieved from \code{strain})}

\item{\code{strain}}{The \code{Strain} object that is being sampled (or a R6Class that inherits from Strain)}

\item{\code{host}}{The \code{Host} object that is carrying the \code{strain} (or a R6Class that inherits from Host)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{WGSequence} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WGSequence-genome"></a>}}
\if{latex}{\out{\hypertarget{method-WGSequence-genome}{}}}
\subsection{Method \code{genome()}}{
Get genome stored in this wgsequence
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WGSequence$genome(as_character = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{as_character}}{Logical indicating if the genome should be returned as a character or integer vector.
Default is TRUE.
'FALSE' returns an integer vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character or integer vector. If an integer vector is specified then 1 = "A", 2 = "C", 3 = "G", and 4 = "T" if genome is DNA based or "U" if genome is RNA based.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WGSequence-print"></a>}}
\if{latex}{\out{\hypertarget{method-WGSequence-print}{}}}
\subsection{Method \code{print()}}{
Print a description of the \code{WGSequence} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WGSequence$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. Description is printed to the console
}
}
}
