% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-intervals.R
\name{plot_intervals}
\alias{plot_intervals}
\title{Plot a histogram of the desired interval in an outbreak}
\usage{
plot_intervals(
  x,
  interval = "transmission",
  show_distribution = FALSE,
  combine = FALSE,
  population_labels = NULL
)
}
\arguments{
\item{x}{An Outbreak or Population object that has had it's outbreak simulated}

\item{interval}{The interval that you want plotted (default is the transmision interval)}

\item{show_distribution}{Logical. If \code{TRUE} and parameters were provided for the interval being plotted then the expected distribution will be overlaid on top of the histogram.}

\item{combine}{A boolean indicating if bins for different populations should be combined (only has an effect if an Outbreak object with multiple populations is provided). If combine is \code{TRUE} and \code{x} is an Outbreak object then \code{show_distribution} will be overridden and set to \code{FALSE}.}

\item{population_labels}{A vector of strings used to label multiple populations. Should be in the same order as the populations in the Outbreak object. (similarly to combine, this only has an effect if an outbreak with multiple populations is provided)}
}
\value{
A ggplot object
}
\description{
Plot a histogram of the desired interval in an outbreak
}
\examples{
# An SEIR model outbreak in a single homogenous population using the 
# serial interval method to find infectors
ref_strain <- ReferenceStrain$new("ref_strain")
population <- Population$new(
     1, 
     ref_strain, 
     inf_rate = 0.75, 
     find_infector_method = "serial", 
     si_shape = 6, si_rate = 1
)
population$run_simulation(Lab$new())
figure <- plot_intervals(population, interval = "serial", show_distribution = TRUE)
plot(figure)

# set.seed(1)
# An SEIR model outbreak with two populations using the tranmission method 
# for finding infectors.
# The two populations will have different transmission interval shapes
outbreak <- Outbreak$new(
    c(
        Population$new(1, ref_strain, init_inf = 5, init_sus = 95),
        Population$new(2, ref_strain, init_inf = 0, init_sus = 100, trans_int_shape = 5)
    ),
    matrix(c(0.75, 0.5, 0.25, 0.1), ncol = 2),
    Lab$new()
)
outbreak$run_simulation()
fig_1 <- plot_intervals(
     outbreak, 
     show_distribution = TRUE, population_labels = c("Canine", "Feline")
)
plot(fig_1)

fig_2 <- plot_intervals(outbreak, combine = TRUE, population_labels = c("Canine", "Feline"))
plot(fig_2)

fig_3 <- plot_intervals(
     outbreak, 
     show_distribution = TRUE, interval = "generation", population_labels = c("Canine", "Feline")
)
plot(fig_3)

}
