% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-population.R
\name{Population}
\alias{Population}
\title{Population}
\description{
An R6 class representing an population. Used to simulate an disease outbreak in one or more
}
\examples{
# An SEIR model outbreak ina population with two groups
set.seed(1)
ref_strain <- ReferenceStrain$new("ref_strain")
population <- Population$new(
    c(
        Group$new(1, ref_strain, init_inf = 5, init_sus = 95),
        Group$new(2, ref_strain, init_inf = 0, init_sus = 100)
    ),
    matrix(c(0.75, 0.5, 0.25, 0.1), ncol = 2),
    Lab$new()
)
population$run_simulation()
# plot each group separately
plot_compartment_dynamics(population, group_labels = c("Senior", "Adult"))
# plot groups combined
plot_compartment_dynamics(population, combine = TRUE)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{inf_rates}}{A matrix with the same number of rows and columns as there is groups. The entry in the \code{i}th row and \code{j}the column is how much infectious pressure an individual in group \code{i} exerts on an individual in group \code{j}}

\item{\code{lab}}{The \code{Lab} object (or an R6Class that inherits from \code{Lab}) that takes samples from \code{groups}}

\item{\code{groups}}{A vector of \code{Group} objects (or R6Classes that inherit from \code{Group})}

\item{\code{time}}{An integer representing the current time in the population}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{exposed_sizes}}{The number of hosts in the exposed compartment in each group}

\item{\code{hosts_due_for_sampling}}{A vector of hosts from all groups that are due sampling at the current time}

\item{\code{is_outbreak_active}}{Logical indicating if there is any host in any group that is still in the exposed or infectious compartment}

\item{\code{infectious_sizes}}{The number of hosts in the infectious compartment in each group}

\item{\code{recovered_sizes}}{The number of hosts in the recovered compartment in each group}

\item{\code{susceptible_sizes}}{The number of hosts in the susceptible compartment in each group}

\item{\code{num_wgs}}{The number of whole genome sequences collected so far}

\item{\code{outbreak_size}}{The number of non_susceptible hosts across all groups}

\item{\code{group_outbreak_sizes}}{The number of non_susceptible hosts in each group}

\item{\code{group_sizes}}{The size of each group}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Population-new}{\code{Population$new()}}
\item \href{#method-Population-get_infectees}{\code{Population$get_infectees()}}
\item \href{#method-Population-get_infectors}{\code{Population$get_infectors()}}
\item \href{#method-Population-get_group_of_infectors}{\code{Population$get_group_of_infectors()}}
\item \href{#method-Population-get_num_infectees}{\code{Population$get_num_infectees()}}
\item \href{#method-Population-infect}{\code{Population$infect()}}
\item \href{#method-Population-print}{\code{Population$print()}}
\item \href{#method-Population-run_simulation}{\code{Population$run_simulation()}}
\item \href{#method-Population-sample_groups}{\code{Population$sample_groups()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-new"></a>}}
\if{latex}{\out{\hypertarget{method-Population-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Population} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$new(groups, inf_rates, lab)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{groups}}{A vector of \code{Group} objects (or R6Classes that inherit from \code{Group})}

\item{\code{inf_rates}}{A matrix with the same number of rows and columns as there is groups. The entry in the \code{i}th row and \code{j}the column is how much infectious pressure an individual in group \code{i} exerts on an individual in group \code{j}}

\item{\code{lab}}{A \code{Lab} object (or R6Class that inherits from \code{Lab})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Population} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-get_infectees"></a>}}
\if{latex}{\out{\hypertarget{method-Population-get_infectees}{}}}
\subsection{Method \code{get_infectees()}}{
Select \code{host}s from the susceptible hosts in each \code{groups} by calling \code{get_infectees()} for each
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$get_infectees(num_infectees)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_infectees}}{A vector of integers indicating how many infectees to select from each group}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of vectors of \code{Host} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-get_infectors"></a>}}
\if{latex}{\out{\hypertarget{method-Population-get_infectors}{}}}
\subsection{Method \code{get_infectors()}}{
Select infectors from the infectious hosts in each group in \code{groups} by calling \code{get_infectors()} for each
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$get_infectors(group_of_infectors, infector_intervals)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{group_of_infectors}}{A vector of  the group id of each infector.}

\item{\code{infector_intervals}}{A vector of integers indicating how long ago each infector was infected / infectious (depends on the \code{find_infector_method} in each group). If \code{find_infector_method="random"} then it's a vector of zeros.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of \code{Host} objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-get_group_of_infectors"></a>}}
\if{latex}{\out{\hypertarget{method-Population-get_group_of_infectors}{}}}
\subsection{Method \code{get_group_of_infectors()}}{
Select the group that each infector comes from
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$get_group_of_infectors(num_infectees, fois_mat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{num_infectees}}{An vector of integers indicating how many infectees to select from each group}

\item{\code{fois_mat}}{A force of infection matrix of the same shape and form as inf_rates}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of vectors of group ids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-get_num_infectees"></a>}}
\if{latex}{\out{\hypertarget{method-Population-get_num_infectees}{}}}
\subsection{Method \code{get_num_infectees()}}{
Determine number of infectees in each group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$get_num_infectees(fois)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fois}}{A vector of the force of infection experienced by each group}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An vector of integers
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-infect"></a>}}
\if{latex}{\out{\hypertarget{method-Population-infect}{}}}
\subsection{Method \code{infect()}}{
Trigger an infection cycle in all \code{groups}. All new infectees (if any) are selected, their infectors are selected and infections of hosts are carried out.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$infect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
This \code{Population} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-print"></a>}}
\if{latex}{\out{\hypertarget{method-Population-print}{}}}
\subsection{Method \code{print()}}{
Print a description of this \code{Population} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$print()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
No return value. Description is printed to the console
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-run_simulation"></a>}}
\if{latex}{\out{\hypertarget{method-Population-run_simulation}{}}}
\subsection{Method \code{run_simulation()}}{
Run this simulation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$run_simulation(feedback = 5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feedback}}{How frequently to print group summary to the console (default 10)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This \code{Population} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Population-sample_groups"></a>}}
\if{latex}{\out{\hypertarget{method-Population-sample_groups}{}}}
\subsection{Method \code{sample_groups()}}{
Collect samples from \code{host}s in every group that are due for sampling
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Population$sample_groups()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
This \code{Population} object
}
}
}
